Pointcloud painter separated from the base class painter3d
The painter3d class is now an abstract class implementing only
the model rotation, translation and scalling.
The primitive related features (drawing the pointcloud)
are now in separated file.
