+++ Makefile
@@ -41,6 +41,7 @@ imagio_exe_CPP_SRCS   = imagio/imagio.cpp \
 							imagio/wimgui/dock.cpp \
 							imagio/wimgui/workspace.cpp \
 							imagio/viewers/3d/painter3d.cpp \
+							imagio/viewers/3d/pointcloud_painter.cpp \
 							imagio/viewers/3d/view3d.cpp\
 							imagio/point_cloud/point_cloud_window.cpp
 imagio_exe_RC_SRCS    = # imagio/imagio.rc
