+++ imagio/viewers/3d/painter3d.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#ifndef WIMGUI_3DPAINT_H
-#define WIMGUI_3DPAINT_H
+#ifndef WIMGUI_PAINTER3D_H
+#define WIMGUI_PAINTER3D_H
 
 #include <vector> 
 #include <string>
@@ -33,56 +33,18 @@
 namespace wimgui
 {
 
-struct vertex
-{
-    float position_x;
-    float position_y;
-    float position_z;
-    float color_r;
-    float color_g;
-    float color_b;
-    float size;
-};
-
-
 class view3d;
 
 
 class painter3d
 {
-private:
-	GLuint vertex_buffer;
-	GLuint vertex_array;
-	gltool::program program;
-	vertex* vertices;
-	unsigned int vertex_index = 0;
-
+protected:
 	glm::mat4 temp_model_matrix, model_matrix, transformation_matrix;
 
 	void stop_rotating();
 	void stop_moving();
 
-#define MAX_VERTICES 64000
 public:
-	painter3d()
-	{
-		vertices = new vertex[MAX_VERTICES * sizeof(vertex)];
-		init_painter();
-	}
-	~painter3d() { delete vertices; }
-	void init_painter();
-	unsigned int get_max_vertices() { return MAX_VERTICES; }
-	unsigned int get_max_bytes() { return MAX_VERTICES * sizeof(vertex); }
-	unsigned int get_vertex_index() { return vertex_index; }
-	vertex* get_vertices() { return vertices;  }
-	GLuint get_vertex_buffer() { return vertex_buffer; }
-	GLuint get_vertex_array() { return vertex_array; }
-	gltool::program* get_program() { return &program; }
-	void draw_point(float x, float y, float z, ImColor& color, float vertex_size=5.0f);
-	void draw_point(glm::vec3& position, ImColor& color, float vertex_size=5.0f)
-	{
-		draw_point(position.x, position.y, position.z, color, vertex_size);
-	}
 	void move(float x, float y);
 	void move(float x, float y, float z);
 	void scale(float wheel);
@@ -90,8 +52,9 @@ public:
 	void set_translation(float x, float y, float z);
 	void rotate(float x, float y);
 	void rotate(float x, float y, float z);
-	void clear();
-	void gl_paint(view3d& view);
+	virtual void clear() = 0;
+	virtual void draw() = 0;
+	virtual void gl_paint(view3d& view) = 0;
 };
 
 }
